feat(access-tokens): renamed access token extension from .model to .schema
